refactor: reorganize device-related types and schemas; update imports for consistency
