+++ src/renderer/store/slices/device/types.ts
@@ -1,82 +1,11 @@
+import {
+  baudRateOptions,
+  deviceConfigurationSchema,
+  devicePinSchema,
+  interfaceOptions,
+  staticHostConfigurationSchema,
+} from '@root/types/PLC/devices'
 import z from 'zod'
-
-const baudRateOptions = ['9600', '14400', '19200', '38400', '57600', '115200'] as const
-
-const interfaceOptions = ['Serial', 'Serial 1', 'Serial 2', 'Serial 3'] as const
-
-const staticHostConfigurationSchema = z.object({
-  ipAddress: z.string(), // This should have the format: XXX.XXX.XXX.XXX
-  dns: z.string(), // This should have the format: XXX.XXX.XXX.XXX
-  gateway: z.string(), // This should have the format: XXX.XXX.XXX.XXX
-  subnet: z.string(), // This should have the format: XXX.XXX.XXX.XXX
-})
-
-const MAC_ADDRESS_REGEX = /^([0-9A-Fa-f]{2})([:\-,])(?:[0-9A-Fa-f]{2}\2){4}[0-9A-Fa-f]{2}$|^[0-9A-Fa-f]{12}$/
-
-type StaticHostConfiguration = z.infer<typeof staticHostConfigurationSchema>
-
-const deviceConfigurationSchema = z.object({
-  deviceBoard: z.string(), // Can be one from the list of boards in the hals file.
-  communicationPort: z.string(),
-  communicationConfiguration: z.object({
-    modbusRTU: z.object({
-      rtuInterface: z.enum(interfaceOptions), // This will be an enumerated that will be associated with the device board selected - Validation will be added further.
-      rtuBaudRate: z.enum(baudRateOptions), // This will be an enumerated that will be associated with the device board selected - Validation will be added further.
-      rtuSlaveId: z.number().int().positive().lte(255).nullable(), // Can be any integer number from 0 to 255 - Validation will be added further.
-      rtuRS485ENPin: z.string().nullable(), // Can be any integer number from 0 to 255 - Validation will be added further.Í
-    }),
-    modbusTCP: z.discriminatedUnion('tcpInterface', [
-      z.object({
-        tcpInterface: z.literal('Wi-Fi'),
-        tcpMacAddress: z.string().regex(MAC_ADDRESS_REGEX).nullable(), // This should have the format: XX:XX:XX:XX:XX:XX
-        tcpWifiSSID: z.string().nullable(),
-        tcpWifiPassword: z.string().nullable(),
-        tcpStaticHostConfiguration: staticHostConfigurationSchema, // When this is omitted the user has chosen DHCP.
-      }),
-      z.object({
-        tcpInterface: z.literal('Ethernet'),
-        tcpMacAddress: z.string().regex(MAC_ADDRESS_REGEX).nullable(),
-        tcpStaticHostConfiguration: staticHostConfigurationSchema, // When this is omitted the user has chosen DHCP.
-      }),
-    ]),
-    communicationPreferences: z.object({
-      enabledRTU: z.boolean(),
-      enabledTCP: z.boolean(),
-      enabledDHCP: z.boolean(),
-    }),
-  }),
-})
-
-type DeviceConfiguration = z.infer<typeof deviceConfigurationSchema>
-
-const pinTypes = ['digitalInput', 'digitalOutput', 'analogInput', 'analogOutput'] as const
-
-type PinTypes = (typeof pinTypes)[number]
-
-/**
- * TODO: Must be filled with the infos that comes from the hals file
- */
-const _defaultPins = []
-/**
- * The pin address obey the following name rules and is populated automatically by the editor.
- *
- * 1. For digital types:
- *    - The address must start with the prefix "%QX" or "%IX"
- *    - Following the prefix, the address must have a integer number starting with 0
- *    - Following the number, the address must have a dot "."
- *    - Following the dot, the address must have a integer number starting with 0 and ending with 7
- * 2. For analog types:
- *    - The address must start with the prefix "%QW" or "%IW"
- *    - Following the prefix, the address must have a integer number starting with 0
- */
-const devicePinSchema = z.object({
-  pin: z.string().max(6),
-  pinType: z.enum(pinTypes),
-  address: z.string(),
-  name: z.string().optional(),
-})
-
-type DevicePin = z.infer<typeof devicePinSchema>
 
 /**
  * The pin mapping is an unique structure that record the pins added by the user.
@@ -195,18 +124,7 @@ type DeviceSlice = DeviceState & {
   deviceActions: DeviceActions
 }
 
-export type {
-  AvailableBoardInfo,
-  DeviceActions,
-  DeviceAvailableOptions,
-  DeviceConfiguration,
-  DevicePin,
-  DevicePinMapping,
-  DeviceSlice,
-  DeviceState,
-  PinTypes,
-  StaticHostConfiguration,
-}
+export type { AvailableBoardInfo, DeviceActions, DeviceAvailableOptions, DevicePinMapping, DeviceSlice, DeviceState }
 export {
   baudRateOptions,
   deviceActionSchema,
@@ -216,5 +134,4 @@ export {
   devicePinSchema,
   deviceStateSchema,
   interfaceOptions,
-  pinTypes,
 }
