+++ src/types/PLC/devices/index.ts
@@ -1,0 +1,2 @@
+export * from './configuration'
+export * from './pin'
